No more nice things.
REMOVED all of the import files that where causing the issues. Removed the 2nd game asset folder that existed with duplicates of EVERYTHING FOR WHATEVER FUCKING REASON.

Created a 2d node with a sprite that I overlayed one of the textures onto as proof that they fucntion. Saved that in a new folder that is called objects.

All of the import files are to my knowledge being made by the asprite wisard or importer or wahtever the fuck its called. It seams to shit itself whenever you import a .aseprite file is acquired from github.